rough-cut of working mysql driver, needs cleaning up

+++ inc/dbic++/param.h
@@ -11,7 +11,7 @@ namespace dbi {
     Param PARAM(string &s);
     Param PARAM(const char* s);
     Param PARAM(const dbi::null &e);
-    Param PARAM_BINARY(unsigned char* data, int l);
+    Param PARAM_BINARY(unsigned char* data, unsigned long l);
 
     ostream& operator<<(ostream &out, Param &p);
 }
